+++ db/updates/2.4.0.php
@@ -1,0 +1,2 @@
+<?php
+file_put_contents(CONNECTOR_DIR.'/db/version', $version);