+++ db/updates/2.4.php
@@ -1,2 +1,0 @@
-<?php
-file_put_contents(CONNECTOR_DIR.'/db/version', $updateFile->getBasename('.php'));