+++ db/updates/2.9.4.php
@@ -2,4 +2,4 @@
 
 $db->query("ALTER TABLE jtl_connector_link_product MODIFY endpoint_id VARCHAR(255)");
 
-file_put_contents(CONNECTOR_DIR.'/db/version', $updateFile->getBasename('.php'));
+file_put_contents(CONNECTOR_DIR.'/db/version', $version);