+++ src/Mapper/PrimaryKeyMapper.php
@@ -1,6 +1,8 @@
 <?php
+
 namespace jtl\Connector\Gambio\Mapper;
 
+use jtl\Connector\Linker\IdentityLinker;
 use jtl\Connector\Mapper\IPrimaryKeyMapper;
 use jtl\Connector\Core\Database\Mysql;
 use jtl\Connector\Core\Logger\Logger;
@@ -10,18 +12,18 @@ class PrimaryKeyMapper implements IPrimaryKeyMapper
     protected $db;
 
     protected static $types = array(
-        1 => 'category',
-        2 => 'customer',
-        4 => 'customer_order',
-        8 => 'delivery_note',
-        16 => 'image',
-        32 => 'manufacturer',
-        64 => 'product',
-        128 => 'specific',
-        256 => 'specific_value',
-        512 => 'payment',
-        1024 => 'crossselling',
-        2048 => 'crossselling_group'
+        IdentityLinker::TYPE_CATEGORY => 'category',
+        IdentityLinker::TYPE_CUSTOMER => 'customer',
+        IdentityLinker::TYPE_CUSTOMER_ORDER => 'customer_order',
+        IdentityLinker::TYPE_DELIVERY_NOTE => 'delivery_note',
+        IdentityLinker::TYPE_IMAGE => 'image',
+        IdentityLinker::TYPE_MANUFACTURER => 'manufacturer',
+        IdentityLinker::TYPE_PRODUCT => 'product',
+        IdentityLinker::TYPE_SPECIFIC => 'specific',
+        IdentityLinker::TYPE_SPECIFIC_VALUE => 'specific_value',
+        IdentityLinker::TYPE_PAYMENT => 'payment',
+        IdentityLinker::TYPE_CROSSSELLING => 'crossselling',
+        IdentityLinker::TYPE_CROSSSELLING_GROUP => 'crossselling_group'
     );
 
     public function __construct()
@@ -33,11 +35,11 @@ class PrimaryKeyMapper implements IPrimaryKeyMapper
     {
         if (isset(static::$types[$type])) {
             $dbResult = $this->db->query("SELECT host_id FROM jtl_connector_link_" . static::$types[$type] . " WHERE endpoint_id = '" . $endpointId . "'");
-    
+
             $host_id = (count($dbResult) > 0) ? $dbResult[0]['host_id'] : null;
-    
+
             Logger::write(sprintf('Trying to get host_id with endpoint_id (%s) and type (%s) ... host_id: (%s)', $endpointId, $type, $host_id), Logger::DEBUG, 'linker');
-    
+
             return $host_id;
         }
     }
@@ -46,11 +48,11 @@ class PrimaryKeyMapper implements IPrimaryKeyMapper
     {
         if (isset(static::$types[$type])) {
             $dbResult = $this->db->query("SELECT endpoint_id FROM jtl_connector_link_" . static::$types[$type] . " WHERE host_id = " . $hostId);
-    
+
             $endpoint_id = (count($dbResult) > 0) ? $dbResult[0]['endpoint_id'] : null;
-    
+
             Logger::write(sprintf('Trying to get endpoint_id with host_id (%s) and type (%s) ... endpoint_id: (%s)', $hostId, $type, $endpoint_id), Logger::DEBUG, 'linker');
-    
+
             return (string)$endpoint_id;
         }
     }
@@ -59,7 +61,7 @@ class PrimaryKeyMapper implements IPrimaryKeyMapper
     {
         if (isset(static::$types[$type])) {
             Logger::write(sprintf('Save link with endpoint_id (%s), host_id (%s) and type (%s)', $endpointId, $hostId, $type), Logger::DEBUG, 'linker');
-    
+
             $this->db->query("INSERT IGNORE INTO jtl_connector_link_" . static::$types[$type] . " (endpoint_id, host_id) VALUES ('" . $endpointId . "'," . $hostId . ")");
         }
     }
@@ -68,17 +70,17 @@ class PrimaryKeyMapper implements IPrimaryKeyMapper
     {
         if (isset(static::$types[$type])) {
             Logger::write(sprintf('Delete link with endpoint_id (%s), host_id (%s) and type (%s)', $endpointId, $hostId, $type), Logger::DEBUG, 'linker');
-    
+
             $where = '';
-    
+
             if ($endpointId && $endpointId != '') {
                 $where = 'endpoint_id = "' . $endpointId . '"';
             }
-    
+
             if ($hostId) {
                 $where = 'host_id = ' . $hostId;
             }
-    
+
             $this->db->query('DELETE FROM jtl_connector_link_' . static::$types[$type] . ' WHERE ' . $where);
         }
     }
@@ -87,8 +89,8 @@ class PrimaryKeyMapper implements IPrimaryKeyMapper
     {
         Logger::write('Clearing linking tables', Logger::DEBUG, 'linker');
 
-        foreach(static::$types as $id => $name) {
-            $this->db->query('TRUNCATE TABLE jtl_connector_link_'.$name);
+        foreach (static::$types as $id => $name) {
+            $this->db->query('TRUNCATE TABLE jtl_connector_link_' . $name);
         }
 
         return true;